review: Put requantization tests in new file, deleted uneeded tests, tested qtype equivalence
